Converted Pyramid master-elements to use Kokkos views.
Was able to re-use a couple of functions that the Tet uses,
and also the hex_volume_grandy function.
