+++ reg_tests/test_files/dgNonConformalElemCylinder/dgNonConformalElemCylinder.norm.gold
@@ -1,10 +1,10 @@
-0.0166514990188203 1 0.002
-0.01242397356257974 2 0.004
-0.004466230783782648 3 0.006
-0.001784512198432304 4 0.008
-0.001150818246964663 5 0.01
-0.0008221735650733199 6 0.012
-0.0006522883084481665 7 0.014
-0.0005852593264633527 8 0.016
-0.0005787430917263411 9 0.018
-0.000603712497841584 10 0.02
+0.01665149901883057 1 0.002
+0.01242397356257037 2 0.004
+0.004466230783784489 3 0.006
+0.001784512198429633 4 0.008
+0.001150818246962169 5 0.01
+0.0008221735650729868 6 0.012
+0.0006522883084471745 7 0.014
+0.000585259326462558 8 0.016
+0.0005787430917263046 9 0.018
+0.0006037124978411584 10 0.02
