+++ reg_tests/test_files/elemHybridFluids/elemHybridFluids.norm.gold
@@ -1,30 +1,30 @@
-0.002469717917465572 1 0.00833333
-0.001860453431738295 2 0.0152778
-0.001394964451773729 3 0.0210648
-0.001145025561142919 4 0.0258873
-0.001026249042757221 5 0.0299061
-0.0009800125651940668 6 0.0332551
-0.0009598664439705263 7 0.0360794
-0.0009640336422932512 8 0.0389037
-0.0009578397848256422 9 0.041728
-0.000948313238279139 10 0.0445523
-0.0009311005605631334 11 0.0473766
-0.0009207481603764571 12 0.0502009
-0.000916598464279311 13 0.0530252
-0.0009194591132344294 14 0.0558495
-0.0009342601555156095 15 0.0586738
-0.000949942120935356 16 0.0614981
-0.0009650804189082575 17 0.0643224
-0.0009910968249845613 18 0.0671467
-0.001017114650456237 19 0.069971
-0.001050205688258983 20 0.0727952
-0.001084270635977166 21 0.0756195
-0.001103390307148286 22 0.0784438
-0.001115502891628467 23 0.0812681
-0.001124855144843051 24 0.0840924
-0.001117640103473386 25 0.0869167
-0.001104757402381038 26 0.089741
-0.0010964784860639 27 0.0925653
-0.001083116086657634 28 0.0953896
-0.001068386313784816 29 0.0982139
-0.001148953357280292 30 0.101038
+0.00246971791746556 1 0.00833333
+0.001860453431738134 2 0.0152778
+0.001394964451773674 3 0.0210648
+0.001145025561142858 4 0.0258873
+0.001026249042757278 5 0.0299061
+0.0009800125651937323 6 0.0332551
+0.0009598664439697512 7 0.0360794
+0.0009640336422915911 8 0.0389037
+0.0009578397848244735 9 0.041728
+0.0009483132382775516 10 0.0445523
+0.0009311005605627318 11 0.0473766
+0.0009207481603785644 12 0.0502009
+0.000916598464280384 13 0.0530252
+0.0009194591132345075 14 0.0558495
+0.0009342601555160624 15 0.0586738
+0.0009499421209351287 16 0.0614981
+0.000965080418909428 17 0.0643224
+0.000991096824983502 18 0.0671467
+0.001017114650455789 19 0.069971
+0.001050205688257602 20 0.0727952
+0.001084270635977932 21 0.0756195
+0.001103390307149772 22 0.0784438
+0.001115502891627346 23 0.0812681
+0.001124855144845239 24 0.0840924
+0.001117640103475439 25 0.0869167
+0.00110475740238066 26 0.089741
+0.001096478486064813 27 0.0925653
+0.001083116086654921 28 0.0953896
+0.001068386313783872 29 0.0982139
+0.001148953357279449 30 0.101038
