+++ src/master_element/MasterElementFactory.C
@@ -11,6 +11,7 @@
 #include <master_element/Hex8CVFEM.h>
 #include <master_element/Hex27CVFEM.h>
 #include <master_element/Tet4CVFEM.h>
+#include <master_element/Pyr5CVFEM.h>
 #include <master_element/MasterElementHO.h>
 
 #include <NaluEnv.h>
